removing the relativePath from pom
