+++ integration-tests/integration-tests-quarkus/pom.xml
@@ -6,7 +6,6 @@
     <groupId>org.kie.kogito</groupId>
     <artifactId>integration-tests</artifactId>
     <version>1.0.0-SNAPSHOT</version>
-    <relativePath>../pom.xml</relativePath>
   </parent>
   <artifactId>integration-tests-quarkus</artifactId>
   <name>Kogito Apps :: Integration Tests :: Quarkus</name>
