--- integration-tests/integration-tests-springboot/src/main/resources/application.properties
@@ -1,26 +1,0 @@
-##
-#  Copyright 2020 Red Hat, Inc. and/or its affiliates.
-#
-#  Licensed under the Apache License, Version 2.0 (the "License");
-#  you may not use this file except in compliance with the License.
-#  You may obtain a copy of the License at
-#
-#        http://www.apache.org/licenses/LICENSE-2.0
-#
-#  Unless required by applicable law or agreed to in writing, software
-#  distributed under the License is distributed on an "AS IS" BASIS,
-#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-#  See the License for the specific language governing permissions and
-#  limitations under the License.
-#
-
-# Runtime Configuration #
-kogito.service.url=${SERVICE_URL:http://localhost:8080}
-kogito.jobs-service.url=http://localhost:8085
-
-spring.mvc.servlet.path=/docs
-
-resteasy.jaxrs.scan-packages=org.kie.kogito.**
-
-server.port=8081
-server.address=0.0.0.0
