+++ integration-tests/integration-tests-springboot/src/test/resources/application.properties
@@ -15,12 +15,8 @@
 #
 
 # Runtime Configuration #
-kogito.service.url=http://localhost:8081
-kogito.jobs-service.url=http://localhost:8085
-
 spring.mvc.servlet.path=/docs
 
 resteasy.jaxrs.scan-packages=org.kie.kogito.**
 
-server.port=8081
 server.address=0.0.0.0
