+++ integration-tests/pom.xml
@@ -4,7 +4,6 @@
     <groupId>org.kie.kogito</groupId>
     <artifactId>kogito-apps</artifactId>
     <version>1.0.0-SNAPSHOT</version>
-    <relativePath>../pom.xml</relativePath>
   </parent>
   <artifactId>integration-tests</artifactId>
   <name>Kogito Apps :: Integration Tests</name>
